modified:   .vscode/Hw/Hw1/cop1 modified:   .vscode/Hw/Hw1/cop1.cpp modified:   .vscode/Hw/Hw1/hw1 modified:   .vscode/Hw/Hw1/hw1.cpp new file:   .vscode/settings.json

+++ .vscode/Hw/Hw1/hw1.cpp
@@ -9,7 +9,7 @@ int main(void) {
     cin >> a >> b >> c;
     cout << "a = " << a << ", b = " << b << ", c = " << c << endl;
     cout << "a*b = " << a*b << endl;
-    cout << "a/b = " << static_cast<float>(a)/b << endl;
+    cout << "a/b = " << (float)a/(float)b << endl;
     cout << "a+b+c = " << a+b+c << endl;
     return 0;
 }
